+++ src/conf.c
@@ -1645,6 +1645,7 @@ void config_setdefaultsettings(Configuration *i)
 
 	safe_strdup(i->oper_snomask, OPER_SNOMASKS);
 	i->server_notice_colors = 1;
+	i->server_notice_show_event = 1;
 	i->ident_read_timeout = 7;
 	i->ident_connect_timeout = 3;
 	i->ban_version_tkl_time = 86400; /* 1d */
@@ -3889,7 +3890,9 @@ int	_conf_oper(ConfigFile *conf, ConfigEntry *ce)
 	oper =  safe_alloc(sizeof(ConfigItem_oper));
 	safe_strdup(oper->name, ce->value);
 
-	oper->server_notice_colors = tempiConf.server_notice_colors; /* default */
+	/* Inherit some defaults: */
+	oper->server_notice_colors = tempiConf.server_notice_colors;
+	oper->server_notice_show_event = tempiConf.server_notice_show_event;
 
 	for (cep = ce->items; cep; cep = cep->next)
 	{
@@ -3936,6 +3939,10 @@ int	_conf_oper(ConfigFile *conf, ConfigEntry *ce)
 		else if (!strcmp(cep->name, "server-notice-colors"))
 		{
 			oper->server_notice_colors = config_checkval(cep->value, CFG_YESNO);
+		}
+		else if (!strcmp(cep->name, "server-notice-show-event"))
+		{
+			oper->server_notice_show_event = config_checkval(cep->value, CFG_YESNO);
 		}
 		else if (!strcmp(cep->name, "modes"))
 		{
@@ -4067,6 +4074,9 @@ int	_test_oper(ConfigFile *conf, ConfigEntry *ce)
 				has_snomask = 1;
 			}
 			else if (!strcmp(cep->name, "server-notice-colors"))
+			{
+			}
+			else if (!strcmp(cep->name, "server-notice-show-event"))
 			{
 			}
 			/* oper::modes */
@@ -7167,6 +7177,9 @@ int	_conf_set(ConfigFile *conf, ConfigEntry *ce)
 		else if (!strcmp(cep->name, "server-notice-colors")) {
 			tempiConf.server_notice_colors = config_checkval(cep->value, CFG_YESNO);
 		}
+		else if (!strcmp(cep->name, "server-notice-show-event")) {
+			tempiConf.server_notice_show_event = config_checkval(cep->value, CFG_YESNO);
+		}
 		else if (!strcmp(cep->name, "level-on-join")) {
 			const char *res = channellevel_to_string(cep->value); /* 'halfop', etc */
 			if (!res)
@@ -7845,6 +7858,9 @@ int	_test_set(ConfigFile *conf, ConfigEntry *ce)
 			}
 		}
 		else if (!strcmp(cep->name, "server-notice-colors")) {
+			CheckNull(cep);
+		}
+		else if (!strcmp(cep->name, "server-notice-show-event")) {
 			CheckNull(cep);
 		}
 		else if (!strcmp(cep->name, "level-on-join")) {